# optimize(be/category): Remove unused image and jig counts