--- a/backend/api/query/category/get_subtree.sql
+++ b/backend/api/query/category/get_subtree.sql
@@ -15,11 +15,6 @@ select distinct id as "id!",
        name,
        created_at,
        updated_at,
-       (select count(*) from image_category where category_id = id)::int8 as "image_count!",
-       (select count(distinct jig.id)::int8
-        from jig
-                 join jig_data_category on (jig.live_id = jig_data_id or jig.draft_id = jig_data_id)
-        where category_id = id)                                           as "jig_count!",
        user_scopes
 
 from path