… backend/api/sqlx-data.json
… backend/api/tests/integration/snapshots/integration__category__delete.snap
… backend/api/tests/integration/snapshots/integration__category__get.snap
… backend/api/tests/integration/snapshots/integration__category__nested_ancestors.snap
… backend/api/tests/integration/snapshots/integration__category__nested_exact.snap
… backend/api/tests/integration/snapshots/integration__category__nested_overlapping.snap
… backend/api/tests/integration/snapshots/integration__category__nested_top_level.snap
… backend/api/tests/integration/snapshots/integration__category__nested_whole_tree.snap
… backend/api/tests/integration/snapshots/integration__category__update_move.snap
… backend/api/tests/integration/snapshots/integration__category__update_parent.snap
… backend/api/tests/integration/snapshots/integration__category__update_rename.snap
… backend/api/tests/integration/snapshots/integration__category__update_reparent_move.snap
… backend/api/tests/integration/snapshots/integration__category__update_scope.snap
… backend/api/tests/integration/snapshots/integration__category__upgdate_ordering-2.snap
… backend/api/tests/integration/snapshots/integration__category__upgdate_ordering.snap