- Removes the image_count and jig_count fields from Category which is not used anywhere in the system.

#### Release trace

![image](https://user-images.githubusercontent.com/4161106/159418218-3aca141c-50a5-4093-8cce-84eed5372044.png)

### `explain analyze` on release database

#### Before

```
 Index Scan using category_index_parent_id_key on category  (cost=0.27..38543.94 rows=331 width=110) (actual time=0.557..155.846 rows=331 loops=1)
   SubPlan 1
     ->  Aggregate  (cost=109.33..109.34 rows=1 width=8) (actual time=0.444..0.444 rows=1 loops=331)
           ->  Seq Scan on image_category  (cost=0.00..109.26 rows=27 width=0) (actual time=0.262..0.442 rows=15 loops=331)
                 Filter: (category_id = category.id)
                 Rows Removed by Filter: 5096
   SubPlan 2
     ->  Aggregate  (cost=6.94..6.95 rows=1 width=8) (actual time=0.024..0.024 rows=1 loops=331)
           ->  Seq Scan on jig_data_category  (cost=0.00..6.93 rows=5 width=0) (actual time=0.020..0.022 rows=1 loops=331)
                 Filter: (category_id = category.id)
                 Rows Removed by Filter: 233
 Planning Time: 0.698 ms
 Execution Time: 156.009 ms
(13 rows)
```

#### After

```
 Sort  (cost=24.16..24.99 rows=331 width=94) (actual time=0.279..0.296 rows=331 loops=1)
   Sort Key: index
   Sort Method: quicksort  Memory: 71kB
   ->  Seq Scan on category  (cost=0.00..10.31 rows=331 width=94) (actual time=0.014..0.083 rows=331 loops=1)
 Planning Time: 0.078 ms
 Execution Time: 0.329 ms
(6 rows)
```