@MendyBerger 2022-03-22
I'm not sure why these fields were in there in the first place, but I'm going to merge this since the savings seem huge and if it's needed we can always add them back in